Cleaned up code and directory
