+++ lib/objects/Event.dart
@@ -5,4 +5,8 @@ class Event {
   bool isDone;
 
   Event({this.title, this.date, this.time, this.isDone});
+
+  String toString() {
+    return '{title: ${this.title}, date: ${this.date}, time: ${this.time}, isDone: ${this.isDone}}';
+  }
 }
